requirements.txt
jinja2>=3.1.0
python-docx>=1.1.0
PyYAML>=6.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
pyjwt>=2.8.0
python-dotenv>=1.0.0
